# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Acessilia Toolbox",
        "description": "Deterministic, stateless capability layer for agentic systems.\n\nCapabilities describe what can be done; providers implement it. The toolbox\nexecutes and normalizes, while goals, planning and provider choice stay with\nthe agentic core.",
        "version": "0.1.0"
    },
    "paths": {
...
        "/v1/capabilities/{capability_id}": {
            "get": {
                "tags": [
                    "capabilities"
                ],
                "summary": "Get Capability",
                "operationId": "get_capability_v1_capabilities__capability_id__get",
                "parameters": [
                    {
                        "name": "capability_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Capability Id"
                        }
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Version"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CapabilityDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Conflict"
                    },
                    "415": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Unsupported Media Type"
                    },
                    "502": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Gateway"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    },
                    "504": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Gateway Timeout"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/capabilities/{capability_id}:execute": {
            "post": {
                "tags": [
                    "capabilities"
                ],
                "summary": "Execute Capability",
                "operationId": "execute_capability_v1_capabilities__capability_id__execute_post",
                "parameters": [
                    {
                        "name": "capability_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Capability Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_execute_capability_v1_capabilities__capability_id__execute_post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExecutionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Conflict"
                    },
                    "415": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Unsupported Media Type"
                    },
                    "502": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Gateway"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    },
                    "504": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Gateway Timeout"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/providers/{provider_id}": {
            "get": {
                "tags": [
                    "providers"
                ],
                "summary": "Get Provider",
                "operationId": "get_provider_v1_providers__provider_id__get",
                "parameters": [
                    {
                        "name": "provider_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Provider Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Get Provider V1 Providers  Provider Id  Get"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Conflict"
                    },
                    "415": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Unsupported Media Type"
                    },
                    "502": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Gateway"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Service Unavailable"
                    },
                    "504": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Gateway Timeout"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ArtifactRef": {
                "properties": {
                    "artifact_id": {
                        "type": "string",
                        "title": "Artifact Id"
                    },
                    "media_type": {
                        "type": "string",
                        "title": "Media Type",
                        "default": "application/json"
                    },
                    "size": {
                        "type": "integer",
                        "title": "Size"
                    }
                },
                "type": "object",
                "required": [
                    "artifact_id",
                    "size"
                ],
                "title": "ArtifactRef",
                "description": "Inline result descriptor.\n\nObject storage arrives with the artifact capabilities; until then the\ndocument travels inline and `artifact_id` is its content fingerprint."
            },
            "Body_execute_capability_v1_capabilities__capability_id__execute_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "contentMediaType": "application/octet-stream",
                        "title": "File",
                        "description": "Document to process"
                    },
                    "provider": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Provider"
                    },
                    "parameters": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Parameters",
                        "description": "JSON object"
                    },
                    "language": {
                        "type": "string",
                        "title": "Language",
                        "default": "pt-BR"
                    },
                    "version": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Version"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_execute_capability_v1_capabilities__capability_id__execute_post"
            },
            "CapabilityDetail": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "version": {
                        "type": "integer",
                        "title": "Version"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "input_schema": {
                        "type": "string",
                        "title": "Input Schema"
                    },
                    "output_schema": {
                        "type": "string",
                        "title": "Output Schema"
                    },
                    "media_types": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Media Types"
                    },
                    "deterministic": {
                        "type": "boolean",
                        "title": "Deterministic"
                    },
                    "cacheable": {
                        "type": "boolean",
                        "title": "Cacheable"
                    },
                    "providers": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Providers"
                    },
                    "requires": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Requires"
                    },
                    "produces": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Produces"
                    },
                    "timeout_hint_seconds": {
                        "type": "integer",
                        "title": "Timeout Hint Seconds"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "version",
                    "description",
                    "input_schema",
                    "output_schema",
                    "media_types",
                    "deterministic",
                    "cacheable",
                    "providers",
                    "requires",
                    "produces",
                    "timeout_hint_seconds"
                ],
                "title": "CapabilityDetail"
            },
...
            "ErrorResponse": {
                "properties": {
                    "code": {
                        "type": "string",
                        "title": "Code"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "details": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Details"
                    }
                },
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "title": "ErrorResponse",
                "description": "Machine-readable error; never carries provider credentials."
            },
            "ExecutionProvenance": {
                "properties": {
                    "capability": {
                        "type": "string",
                        "title": "Capability"
                    },
                    "capability_version": {
                        "type": "integer",
                        "minimum": 1.0,
                        "title": "Capability Version"
                    },
                    "provider": {
                        "type": "string",
                        "title": "Provider"
                    },
                    "provider_version": {
                        "type": "string",
                        "title": "Provider Version"
                    },
                    "input_fingerprints": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Input Fingerprints"
                    },
                    "output_fingerprints": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Output Fingerprints"
                    },
                    "parameters_hash": {
                        "type": "string",
                        "title": "Parameters Hash"
                    },
                    "model_versions": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object",
                        "title": "Model Versions"
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Started At"
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Completed At"
                    },
                    "duration_ms": {
                        "type": "integer",
                        "minimum": 0.0,
                        "title": "Duration Ms"
                    },
                    "cache_key": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Cache Key"
                    },
                    "cache_hit": {
                        "type": "boolean",
                        "title": "Cache Hit",
                        "default": false
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "capability",
                    "capability_version",
                    "provider",
                    "provider_version",
                    "parameters_hash",
                    "started_at",
                    "completed_at",
                    "duration_ms"
                ],
                "title": "ExecutionProvenance",
                "description": "Everything needed to reproduce, audit or invalidate a result."
            },
            "ExecutionResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "capability": {
                        "type": "string",
                        "title": "Capability"
                    },
                    "provider": {
                        "type": "string",
                        "title": "Provider"
                    },
                    "artifacts": {
                        "items": {
                            "$ref": "#/components/schemas/ArtifactRef"
                        },
                        "type": "array",
                        "title": "Artifacts"
                    },
                    "document": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Document"
                    },
                    "provenance": {
                        "$ref": "#/components/schemas/ExecutionProvenance"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "status",
                    "capability",
                    "provider",
                    "artifacts",
                    "document",
                    "provenance"
                ],
                "title": "ExecutionResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}